Made updates and changes with given suggestions in #57
- Removed `cpt_capture_key` because key inputs are captured differently now.
- Set defaults to what they should be given what they were previously
- Put bindings related functions are `bindings.cpp`
- Separated out stuff into a header file `bindings.h`
- Put IMGUI stuff related to the plugin class in its own file `PluginSettings.cpp`

The biggest contentious thing IMO is around `CheckpointPlugin.cpp`L578 where I discriminate against binds that would affect steer and pitch.
