+++ CheckpointPlugin.cpp
@@ -16,10 +16,6 @@
 
 #include "bakkesmod/wrappers/ArrayWrapper.h"
 
-#include "imgui.h"
-#include "imgui_internal.h"
-#include "imgui_rangeslider.h"
-
 using namespace std::placeholders;
 
 const std::string_view DEFAULT_SAVE_FILE_NAME = "freeplaycheckpoint.data";
@@ -417,491 +413,8 @@ void CheckpointPlugin::nextCheckpoint(std::vector<std::string> command) {
     loadCurCheckpoint();
 }
 
-inline void CheckpointPlugin::OnKeyPressed(ActorWrapper aw, void* params, std::string eventName) {
-    s* p = reinterpret_cast<s*>(params);
-    if (p->t.e != 0) { return; }
-
-    ImGui::CloseCurrentPopup();
-    cvarManager->executeCommand("closemenu checkpointplugin", false);
-    gameWrapper->UnhookEvent("Function TAGame.GameViewportClient_TA.HandleKeyPress");
-    gameWrapper->UnhookEvent("Function TAGame.GameViewportClient_TA.HandleAxisPress");
-    std::string key = gameWrapper->GetFNameByIndex(p->k.i);
-
-    // this is because of mem-access issues ... even though they may still exist
-    gameWrapper->Execute([this, key, et = p->t.e](GameWrapper* gw) {
-        cvarManager->getCvar(keys_to_cvars.at(which_is_being_bound)).setValue(key);
-        which_is_being_bound = KEYBIND_ASSIGNWHICH::NONE;
-        });
-}
-
-inline void CheckpointPlugin::OnKeyAxisInput(ActorWrapper aw, void* params, std::string eventName) {
-    s* p = reinterpret_cast<s*>(params);
-
-    // this is done in ways to emulate how the game does it
-    if (std::fabs(p->t.d) <= 0.8) {
-        // commit to a direction, lol
-        return;
-    }
-
-    std::string key = gameWrapper->GetFNameByIndex(p->k.i);
-    if (key.contains("Mouse") && std::abs(p->t.d) < 3) {
-        // a delta of 3 seems to be the threshhold for a mouse axis input
-        return;
-    }
-
-    ImGui::CloseCurrentPopup();
-    cvarManager->executeCommand("closemenu checkpointplugin", false);
-    gameWrapper->UnhookEvent("Function TAGame.GameViewportClient_TA.HandleAxisPress");
-    gameWrapper->UnhookEvent("Function TAGame.GameViewportClient_TA.HandleKeyPress");
-
-    // this is because of mem-access issues ... even though they may still exist
-    gameWrapper->Execute([this, key, d = p->t.d](GameWrapper* gw) {
-        if (key == "XboxTypeS_LeftTriggerAxis" || key == "XboxTypeS_RightTriggerAxis") {
-            // this doesn't have a positive/negative direction, so it's not 2 inputs in one
-            cvarManager->getCvar(keys_to_cvars.at(which_is_being_bound)).setValue(key);
-        } else {
-            std::string first_set = signbit(d) ? "-" : "+";
-            std::string second_set = signbit(d) ? "+" : "-";
-            switch (which_is_being_bound) {
-            case KEYBIND_ASSIGNWHICH::CPT_REWIND_KEY:
-                cvarManager->getCvar(keys_to_cvars.at(KEYBIND_ASSIGNWHICH::CPT_REWIND_KEY)).setValue(key + first_set);
-                cvarManager->getCvar(keys_to_cvars.at(KEYBIND_ASSIGNWHICH::CPT_FASTFORWARD_KEY)).setValue(key + second_set);
-                break;
-            case KEYBIND_ASSIGNWHICH::CPT_FASTFORWARD_KEY:
-                cvarManager->getCvar(keys_to_cvars.at(KEYBIND_ASSIGNWHICH::CPT_FASTFORWARD_KEY)).setValue(key + first_set);
-                cvarManager->getCvar(keys_to_cvars.at(KEYBIND_ASSIGNWHICH::CPT_REWIND_KEY)).setValue(key + second_set);
-                break;
-            }
-
-        }
-        which_is_being_bound = KEYBIND_ASSIGNWHICH::NONE;
-        });
-}
-
-/**
-* @brief
-* @details Done within the IMGUI section.
-*
-*/
-inline void CheckpointPlugin::OpenMenuForKeybinding() {
-    close_opened_menus(cvarManager, gameWrapper);
-    cvarManager->executeCommand("openmenu checkpointplugin", false);
-    gameWrapper->HookEventWithCaller<ActorWrapper>(
-        "Function TAGame.GameViewportClient_TA.HandleKeyPress",
-        std::bind(
-            &CheckpointPlugin::OnKeyPressed,
-            this,
-            std::placeholders::_1,
-            std::placeholders::_2,
-            std::placeholders::_3));
-
-    if (which_is_being_bound == KEYBIND_ASSIGNWHICH::CPT_REWIND_KEY || which_is_being_bound == KEYBIND_ASSIGNWHICH::CPT_FASTFORWARD_KEY) {
-        gameWrapper->HookEventWithCaller<ActorWrapper>(
-            "Function TAGame.GameViewportClient_TA.HandleAxisPress",
-            std::bind(
-                &CheckpointPlugin::OnKeyAxisInput,
-                this,
-                std::placeholders::_1,
-                std::placeholders::_2,
-                std::placeholders::_3));
-    }
-}
-
-void CheckpointPlugin::RenderSettings() {
-    static const auto openmenufunc = [this]() {
-        gameWrapper->Execute([this](GameWrapper* gw) {
-            OpenMenuForKeybinding();
-            });
-        };
-
-    // main driver for rendering plugin settings
-    ImGui::TextUnformatted("Bindings");
-    ImGui::TextUnformatted("Instructions: Click the button corresponding to the binding to choose a button to set.");
-    ImGui::TextUnformatted("After assigning bindings to each action, click \"Apply All Bindings\" to set them.");
-
-    ImGui::Separator();
-    ImGui::PushStyleVar(ImGuiStyleVar_ItemSpacing, ImVec2(0.f, ImGui::GetStyle().ItemSpacing.y));
-
-    ImGui::BeginColumns("bindings columns##bc", 3, ImGuiColumnsFlags_NoBorder | ImGuiColumnsFlags_NoResize);
-    ImGui::SetColumnWidth(0, 280.f);
-    ImGui::SetColumnWidth(1, 230.f);
-    ImGui::SetColumnWidth(2, 200.f);
-
-    // freeze key
-    if (ImGui::Button("Freeze (cpt_freeze)##fz", ImVec2(250.f, 0.f))) {
-        which_is_being_bound = KEYBIND_ASSIGNWHICH::CPT_FREEZE_KEY;
-        openmenufunc();
-    }
-
-    ImGui::NextColumn();
-    static std::string cfk;
-    cfk = cvarManager->getCvar("cpt_freeze_key").getStringValue();
-    ImGui::Text("[ %s ]", cfk.c_str());
-
-    ImGui::NextColumn();
-    ImGui::NextColumn();
-
-    // do checkpoint
-    if (ImGui::Button("Checkpoint (cpt_do_checkpoint)##cp", ImVec2(250.f, 0.f))) {
-        which_is_being_bound = KEYBIND_ASSIGNWHICH::CPT_DO_CHECKPOINT_KEY;
-        openmenufunc();
-    }
-
-    ImGui::NextColumn();
-
-    static std::string cdck;
-    cdck = cvarManager->getCvar("cpt_do_checkpoint_key").getStringValue();
-    ImGui::Text("[ %s ] ", cdck.c_str());
-
-    ImGui::NextColumn();
-    ImGui::NextColumn();
-
-    // prev checkpoint
-    if (ImGui::Button("Prev. Checkpoint (cpt_prev_checkpoint)##pc", ImVec2(250.f, 0.f))) {
-        which_is_being_bound = KEYBIND_ASSIGNWHICH::CPT_PREV_CHECKPOINT_KEY;
-        openmenufunc();
-    }
-
-    ImGui::NextColumn();
-
-    static std::string cpck;
-    cpck = cvarManager->getCvar("cpt_prev_checkpoint_key").getStringValue();
-    ImGui::Text("[ %s ]", cpck.c_str());
-
-    ImGui::NextColumn();
-
-    static bool ignore_while_playing_prev;
-    ignore_while_playing_prev = cvarManager->getCvar("cpt_ignore_prev").getBoolValue();
-    if (ImGui::Checkbox("Ignore While Playing##prev", &ignore_while_playing_prev)) {
-        cvarManager->getCvar("cpt_ignore_prev").setValue(ignore_while_playing_prev);
-    }
-
-    ImGui::NextColumn();
-
-    // next checkpoint
-    if (ImGui::Button("Next Checkpoint (cpt_next_checkpoint)##nc", ImVec2(250.f, 0.f))) {
-        which_is_being_bound = KEYBIND_ASSIGNWHICH::CPT_NEXT_CHECKPOINT_KEY;
-        openmenufunc();
-    }
-
-    ImGui::NextColumn();
-
-    static std::string cnck;
-    cnck = cvarManager->getCvar("cpt_next_checkpoint_key").getStringValue();
-    ImGui::Text("[ %s ]", cnck.c_str());
-
-    ImGui::NextColumn();
-
-    static bool ignore_while_playing_next;
-    ignore_while_playing_next = cvarManager->getCvar("cpt_ignore_next").getBoolValue();
-    if (ImGui::Checkbox("Ignore While Playing##next", &ignore_while_playing_next)) {
-        cvarManager->getCvar("cpt_ignore_next").setValue(ignore_while_playing_next);
-    }
-
-    ImGui::NextColumn();
-
-    // freeze ball
-    if (ImGui::Button("Freeze Ball/Unfreeze Car (cpt_freeze_ball)##fbuc", ImVec2(250.f, 0.f))) {
-        which_is_being_bound = KEYBIND_ASSIGNWHICH::CPT_FREEZE_BALL_KEY;
-        openmenufunc();
-    }
-
-    ImGui::NextColumn();
-
-    static std::string cfbuc;
-    cfbuc = cvarManager->getCvar("cpt_freeze_ball_key").getStringValue();
-    ImGui::Text("[ %s ]", cfbuc.c_str());
-
-    ImGui::NextColumn();
-
-    static bool ignore_freeze_ball;
-    ignore_freeze_ball = cvarManager->getCvar("cpt_ignore_freeze_ball").getBoolValue();
-    if (ImGui::Checkbox("Ignore While Playing##fzbl", &ignore_freeze_ball)) {
-        cvarManager->getCvar("cpt_ignore_freeze_ball").setValue(ignore_freeze_ball);
-    }
-
-    ImGui::NewLine();
-    ImGui::NextColumn();
-
-    // mirror state
-    if (ImGui::Button("Mirror shot (cpt_mirror_state)##ms", ImVec2(250.f, 0.f))) {
-        which_is_being_bound = KEYBIND_ASSIGNWHICH::CPT_MIRROR_STATE_KEY;
-        openmenufunc();
-    }
-
-    ImGui::NextColumn();
-
-    static std::string cmsk;
-    cmsk = cvarManager->getCvar("cpt_mirror_state_key").getStringValue();
-    ImGui::Text("[ %s ]", cmsk.c_str());
-
-    ImGui::NextColumn();    ImGui::NextColumn();
-
-    // rewind
-    if (ImGui::Button("Rewind (cpt_rewind)##rw", ImVec2(250.f, 0.f))) {
-        which_is_being_bound = KEYBIND_ASSIGNWHICH::CPT_REWIND_KEY;
-        openmenufunc();
-    }
-
-    ImGui::NextColumn();
-
-    static std::string crwk;
-    crwk = cvarManager->getCvar("cpt_rewind_key").getStringValue();
-    ImGui::Text("[ %s ]", crwk.c_str());
-
-    ImGui::NextColumn();    ImGui::NextColumn();
-
-    // fast forward
-    if (ImGui::Button("Fast Forward (cpt_fastforward)##ff", ImVec2(250.f, 0.f))) {
-        which_is_being_bound = KEYBIND_ASSIGNWHICH::CPT_FASTFORWARD_KEY;
-        openmenufunc();
-    }
-
-    ImGui::NextColumn();
-
-    static std::string cffk;
-    cffk = cvarManager->getCvar("cpt_fastforward_key").getStringValue();
-    ImGui::Text("[ %s ]", cffk.c_str());
-
-    ImGui::EndColumns();
-    ImGui::PopStyleVar();
-
-    ImGui::NewLine();
-
-    // apply all bindings
-    if (ImGui::Button("Apply All Bindings##aab")) {
-        applyBindKeys({});
-    }
-
-    ImGui::SameLine(0.f, 50.f);
-    ImGui::TextUnformatted("(if bindings have never been set)");
-
-    ImGui::SameLine(0.f, 50.f);
-    // remove binding
-    ImGui::SetNextItemWidth(20.f);
-    if (ImGui::Button("Remove Bindings##rb")) {
-        removeBindKeys({});
-    }
-    ImGui::SameLine(0.f, 10.f);
-    ImGui::TextUnformatted("*Removes action on the button from the adjacent key* also :(");
-
-    // reset default bindings
-    if (ImGui::Button("Reset Default Bindings##rdb")) {
-        resetDefaultBindKeys({});
-    }
-    ImGui::SameLine(0.f, 50.f);
-    ImGui::TextUnformatted("(does not apply them)");
-
-    // disable working while in custom training
-    static bool disable_training;
-    disable_training = cvarManager->getCvar("cpt_disable_training").getBoolValue();
-    if (ImGui::Checkbox("Disable binds in custom training##dbict", &disable_training)) {
-        cvarManager->getCvar("cpt_disable_training").setValue(disable_training);
-    }
-
-    // disable working while in a workshop map
-    static bool disable_workshop;
-    disable_workshop = cvarManager->getCvar("cpt_disable_workshop").getBoolValue();
-    if (ImGui::Checkbox("Disable binds in workshop##dbiw", &disable_workshop)) {
-        cvarManager->getCvar("cpt_disable_workshop").setValue(disable_workshop);
-    }
-
-    ImGui::NewLine();
-
-    // reset button
-    ImGui::TextUnformatted("Reset Shot button loads last checkpoint instead of resetting if loaded before");
-    static int load_after_reset;
-    load_after_reset = cvarManager->getCvar("cpt_load_after_reset").getIntValue();
-    if (ImGui::SliderInt("(seconds)##rb", &load_after_reset, 0, 30)) {
-        cvarManager->getCvar("cpt_load_after_reset").setValue(load_after_reset);
-    }
-
-    ImGui::Separator();
-    ImGui::NewLine();
-
-    // car and ball variance
-    ImGui::TextUnformatted("Variance - applied when leaving rewind mode");
-    ImGui::Separator();
-    static float variance_car_dir;
-    variance_car_dir = cvarManager->getCvar("cpt_variance_car_dir").getFloatValue();
-    if (ImGui::SliderFloat("Car Direction (degrees)##cdd", &variance_car_dir, 0.0f, 30.0f, "%.2f")) {
-        cvarManager->getCvar("cpt_variance_car_dir").setValue(variance_car_dir);
-    }
-
-    static float variance_car_spd;
-    variance_car_spd = cvarManager->getCvar("cpt_variance_car_spd").getFloatValue();
-    if (ImGui::SliderFloat("Car Speed (percent)##css", &variance_car_spd, 0.0f, 50.0f, "%.2f")) {
-        cvarManager->getCvar("cpt_variance_car_spd").setValue(variance_car_spd);
-    }
-
-    static float variance_car_rot[2];
-    static std::vector<std::string> twoval_holder;
-    static std::string var_car_rot;
-    var_car_rot = cvarManager->getCvar("cpt_variance_car_rot").getStringValue();
-    replace(var_car_rot, "(", "");
-    replace(var_car_rot, ")", "");
-    split(var_car_rot, twoval_holder, ',');
-    switch (twoval_holder.size()) {
-    case 0:
-        variance_car_rot[0] = 0.0f;
-        variance_car_rot[1] = 0.0f;
-        break;
-    case 1:
-        variance_car_rot[0] = variance_car_rot[1] = std::stof(twoval_holder[0]);
-        break;
-    case 2:
-        variance_car_rot[0] = std::stof(twoval_holder[0]);
-        variance_car_rot[1] = std::stof(twoval_holder[1]);
-        break;
-    }
-    if (ImGui::RangeSliderFloat("Car Rotation (strength)##crs", &variance_car_rot[0], &variance_car_rot[1], 0.0f, 10.0f, "(%.2f, %.2f)")) {
-        cvarManager->getCvar("cpt_variance_car_rot").setValue(std::format("({}, {})", variance_car_rot[0], variance_car_rot[1]));
-    }
-
-    static float variance_ball_dir;
-    variance_ball_dir = cvarManager->getCvar("cpt_variance_ball_dir").getFloatValue();
-    if (ImGui::SliderFloat("Ball Direction (degrees)##bdd", &variance_ball_dir, 0.0f, 30.0f, "%.2f")) {
-        cvarManager->getCvar("cpt_variance_ball_dir").setValue(variance_ball_dir);
-    }
-
-    static float variance_ball_spd;
-    variance_ball_spd = cvarManager->getCvar("cpt_variance_ball_spd").getFloatValue();
-    if (ImGui::SliderFloat("Ball Speed (percent)##bsp", &variance_ball_spd, 0.0f, 50.0f, "%.2f")) {
-        cvarManager->getCvar("cpt_variance_ball_spd").setValue(variance_ball_spd);
-    }
-
-    static float variance_ball_rot[2];
-    static std::string var_ball_rot;
-    var_ball_rot = cvarManager->getCvar("cpt_variance_ball_rot").getStringValue();
-    replace(var_ball_rot, "(", "");
-    replace(var_ball_rot, ")", "");
-    split(var_ball_rot, twoval_holder, ',');
-    switch (twoval_holder.size()) {
-    case 0:
-        variance_ball_rot[0] = 0.0f;
-        variance_ball_rot[1] = 0.0f;
-        break;
-    case 1:
-        variance_ball_rot[0] = variance_ball_rot[1] = std::stof(twoval_holder[0]);
-        break;
-    case 2:
-        variance_ball_rot[0] = std::stof(twoval_holder[0]);
-        variance_ball_rot[1] = std::stof(twoval_holder[1]);
-        break;
-    }
-    if (ImGui::RangeSliderFloat("Ball Rotation (strength)##brs", &variance_ball_rot[0], &variance_ball_rot[1], 0.0f, 10.0f, "(%.2f, %.2f)")) {
-        cvarManager->getCvar("cpt_variance_ball_rot").setValue(std::format("({}, {})", variance_ball_rot[0], variance_ball_rot[1]));
-    }
-
-    static int variance_tot;
-    variance_tot = cvarManager->getCvar("cpt_variance_tot").getIntValue();
-    if (ImGui::SliderInt("Max. Total Variance##mtv", &variance_tot, 0, 50)) {
-        cvarManager->getCvar("cpt_variance_tot").setValue(variance_tot);
-    }
-
-    ImGui::NewLine();
-    // mirror randomly
-    static bool mirror_random;
-    mirror_random = cvarManager->getCvar("cpt_mirror_loads").getBoolValue();
-    if (ImGui::Checkbox("Randomly mirror when loading checkpoint##mrand", &mirror_random)) {
-        cvarManager->getCvar("cpt_mirror_loads").setValue(mirror_random);
-    }
-
-    // random chkpt
-    static bool checkpoint_random;
-    checkpoint_random = cvarManager->getCvar("cpt_randomize_loads").getBoolValue();
-    if (ImGui::Checkbox("Load random checkpoint instead of latest##crand", &checkpoint_random)) {
-        cvarManager->getCvar("cpt_randomize_loads").setValue(checkpoint_random);
-    }
-    ImGui::Separator();
-
-    // auto-reset checkpoint
-    ImGui::NewLine();
-    ImGui::TextUnformatted("Auto-reset checkpoint - reset when the following occurs:");
-    ImGui::Separator();
-    static bool goal_scored;
-    goal_scored = cvarManager->getCvar("cpt_reset_on_goal").getBoolValue();
-    if (ImGui::Checkbox("Goal Scored", &goal_scored)) {
-        cvarManager->getCvar("cpt_reset_on_goal").setValue(goal_scored);
-    }
-    static bool reset_on_ball_ground;
-    reset_on_ball_ground = cvarManager->getCvar("cpt_reset_on_ball_ground").getBoolValue();
-    if (ImGui::Checkbox("Ball touches ground", &reset_on_ball_ground)) {
-        cvarManager->getCvar("cpt_reset_on_ball_ground").setValue(reset_on_ball_ground);
-    }
-    static bool next_instead_of_reset;
-    next_instead_of_reset = cvarManager->getCvar("cpt_next_instead_of_reset").getBoolValue();
-    if (ImGui::Checkbox("Load next checkpoint instead of resetting", &next_instead_of_reset)) {
-        cvarManager->getCvar("cpt_next_instead_of_reset").setValue(next_instead_of_reset);
-    }
-    ImGui::Separator();
-    ImGui::NewLine();
-
-    // else
-    ImGui::TextUnformatted("Other options:");
-    ImGui::Separator();
-    ImGui::TextUnformatted("Save File Name:");
-    ImGui::SameLine(0.0f, 50.0f);
-
-    static char fn_str[1048] = "";
-    static std::once_flag never_again;
-    std::call_once(never_again, [this](char* str) {
-        std::string filename = cvarManager->getCvar("cpt_filename").getStringValue();
-        std::strncpy(fn_str, filename.c_str(), 1024);
-        }, fn_str);
-    if (ImGui::InputText("##filenametb", fn_str, 1024)) { // maximum 1024 characters
-        cvarManager->getCvar("cpt_filename").setValue(fn_str);
-    }
-    if (ImGui::SmallButton("Delete ALL Shots (even locked shots; not undo-able!)##das")) {
-        deleteAllCheckpoints({});
-    }
-    ImGui::SameLine(0.0f, 50.0f);
-
-    ImGui::NewLine();
-    static bool show_boost;
-    show_boost = cvarManager->getCvar("cpt_show_boost").getBoolValue();
-    if (ImGui::Checkbox("Show player boost while rewinding##sb", &show_boost)) {
-        cvarManager->getCvar("cpt_show_boost").setValue(show_boost);
-    }
-    static bool clean_hist;
-    clean_hist = cvarManager->getCvar("cpt_clean_history").getBoolValue();
-    if (ImGui::Checkbox("Clean History -- Erases future history points when resuming##ch", &clean_hist)) {
-        cvarManager->getCvar("cpt_clean_history").setValue(clean_hist);
-    }
-    static int hist_len;
-    hist_len = cvarManager->getCvar("cpt_history_length").getIntValue();
-    if (ImGui::SliderInt("History Length (seconds)##hlen", &hist_len, 10, 120)) {
-        cvarManager->getCvar("cpt_history_length").setValue(hist_len);
-    }
-    static int hist_refresh;
-    hist_refresh = cvarManager->getCvar("cpt_snapshot_interval").getIntValue();
-    if (ImGui::SliderInt("History Refresh Rate (ms)##hrr", &hist_refresh, 1, 10)) {
-        cvarManager->getCvar("cpt_snapshot_interval").setValue(hist_refresh);
-    }
-    ImGui::NewLine();
-    static bool show_debug;
-    show_debug = cvarManager->getCvar("cpt_debug").getBoolValue();
-    if (ImGui::Checkbox("Debug -- Show debugging state##dbg", &show_debug)) {
-        cvarManager->getCvar("cpt_debug").setValue(show_debug);
-    }
-
-    ImGui::NewLine();
-    ImGui::NewLine();
-
-    ImGui::TextUnformatted("Freeplay Checkpoint");
-    ImGui::TextUnformatted("Bugs/Feature Requests:");
-    TextURL("github.com/NitrOP7674", "https://github.com/NitrOP7674", TRUE, TRUE);
-    ImGui::TextUnformatted(" -or- on Discord:");
-    TextURL("https://discord.gg/SPBxrtfrZw", "https://discord.gg/SPBxrtfrZw", TRUE, FALSE);
-    ImGui::TextUnformatted("** Please make sure to read the README first! **");
-}
-
 std::string CheckpointPlugin::GetPluginName() {
     return PLUGIN_NAME;
-}
-
-void CheckpointPlugin::SetImGuiContext(uintptr_t ctx) {
-    ImGui::SetCurrentContext(reinterpret_cast<ImGuiContext*>(ctx));
 }
 
 void CheckpointPlugin::lockCheckpoint(std::vector<std::string> command) {
@@ -1062,6 +575,41 @@ void CheckpointPlugin::OnPreAsync(std::string funcName) {
     }
 
     if (rewindMode) {
+        ignore_pitch_clearing = false;
+        ignore_steer_clearing = false;
+
+        static const auto removeDashPlusMaybe = [](std::string& key) {
+            if (key.back() == '+' || key.back() == '-') {
+                key.pop_back();
+            }
+            };
+        static const auto removeXYMaybe = [](std::string& key) {
+            if (key.back() == 'X' || key.back() == 'Y') {
+                key.pop_back();
+            }
+            };
+        std::string rewindKey = cvarManager->getCvar("cpt_rewind_key").getStringValue();
+        std::string ffKey = cvarManager->getCvar("cpt_fastforward_key").getStringValue();
+        removeDashPlusMaybe(rewindKey);
+        removeDashPlusMaybe(ffKey);
+        removeXYMaybe(rewindKey);
+        removeXYMaybe(ffKey);
+
+        SettingsWrapper settings;
+        auto bindings = settings.GetAllGamepadBindings();
+        bindings.append_range(settings.GetAllGamepadBindings());
+
+        // Prevent things like XboxTypeS_LeftX / XboxTypeS_LeftY from exiting rewind mode
+        // note: this will also catch XboxTypeS_LeftThumbstick
+        // I think the compromise of not accidentally exiting the mode is cool
+        // It could be a setting boolean, but I don't like the game resuming because of a 
+        // thumbstick input being a thing anyway.
+        for (const auto& bind : bindings){
+            if (bind.first.contains(rewindKey) || bind.first.contains(ffKey)) {
+                if (bind.second.contains("Steer")) { ignore_steer_clearing = true; }
+                if (bind.second.contains("Pitch")) { ignore_pitch_clearing = true; }
+            }
+        }
         // entering rewind mode, so hook getting rewind/ff inputs.
         gameWrapper->HookEventWithCallerPost<ActorWrapper>(
             "Function TAGame.GameViewportClient_TA.HandleKeyPress", [this](ActorWrapper w, void* params, std::string eventName) {
@@ -1091,11 +639,6 @@ void CheckpointPlugin::OnPreAsync(std::string funcName) {
             "Function TAGame.GameViewportClient_TA.HandleAxisPress",
             [this](ActorWrapper w, void* params, std::string eventName) {
                 s* p = reinterpret_cast<s*>(params);
-                static const auto removeLastMaybe = [](std::string& key) {
-                    if (key.back() == '+' || key.back() == '-') {
-                        key.pop_back();
-                    }
-                    };
 
                 static const auto is_neg = [](float f) { return signbit(f) == 1; };
                 static const auto is_pos = [](float f) { return signbit(f) == 0; };
@@ -1105,8 +648,8 @@ void CheckpointPlugin::OnPreAsync(std::string funcName) {
                 std::string rewindKey = cvarManager->getCvar("cpt_rewind_key").getStringValue(), rk2 = rewindKey;
                 std::string ffKey = cvarManager->getCvar("cpt_fastforward_key").getStringValue(), fk2 = ffKey;
                 std::string key = gameWrapper->GetFNameByIndex(p->k.i);
-                removeLastMaybe(rewindKey);
-                removeLastMaybe(ffKey);
+                removeDashPlusMaybe(rewindKey);
+                removeDashPlusMaybe(ffKey);
 
                 if (key == rewindKey || key == ffKey) {
                     static bool catch_once = false;
@@ -1173,6 +716,10 @@ bool CheckpointPlugin::rewind(ServerWrapper sw) {
         (ci.HoldingBoost ? 0x20 : 0) |
         (abs(ci.Steer) >= .05 ? 0x40 : 0) |
         (abs(ci.Pitch) >= .05 ? 0x80 : 0);
+    // if rewind key or ff key affects steer/pitch, clear them from the discriminator
+    if (ignore_steer_clearing) { buttonsDown &= ~(0x40); };
+    if (ignore_pitch_clearing) { buttonsDown &= ~(0x80); };
+
     // See if we should exit rewind mode due to input.
     if (buttonsDown != 0 || rewindState.atCheckpoint || rewindState.justLoadedQuickCheckpoint) {
         if (buttonsDown > rewindState.buttonsDown && currentTime - lastRecordTime > 0.1f) {
@@ -1197,7 +744,6 @@ bool CheckpointPlugin::rewind(ServerWrapper sw) {
         return true; // Staying in rewind; apply state.
     }
     rewindState.buttonsDown = buttonsDown;
-    log(std::format("last seek val: {}", rewindState.lastSeekVal));
     // Determine how much to rewind / advance time.
     rewindState.deleting = false;
     if (rewindState.lastSeekVal < -.95 && rewindState.holdingFor <= 0) {
@@ -1317,6 +863,7 @@ void CheckpointPlugin::Render(CanvasWrapper canvas) {
         size_t current = std::clamp<size_t>(
             history.size() + size_t(ceil(rewindState.virtualTimeOffset / snapshotInterval)),
             0, history.size() - 1);
+        show(canvas, &loc, "lastSeekVal: " + std::to_string(rewindState.lastSeekVal));
         show(canvas, &loc, "current: " + std::to_string(current));
     }
     if (!rewindMode) {
@@ -1405,123 +952,3 @@ void CheckpointPlugin::saveCheckpointFile() {
     }
     out.close();
 }
-
-/**
- * @brief do the following on menu open
- */
-void CheckpointPlugin::OnOpen() {
-    if (which_is_being_bound == KEYBIND_ASSIGNWHICH::NONE) {
-        gameWrapper->Execute(
-            [this](GameWrapper* gw) { cvarManager->executeCommand("closemenu checkpointplugin", false); });
-    }
-};
-
-/**
- * @brief do the following on menu close
- */
-void CheckpointPlugin::OnClose() {
-    gameWrapper->Execute([this](GameWrapper* gw) { reopen_closed_menus(cvarManager); });
-};
-
-/**
- * @brief (ImGui) Code called while rendering your menu window
- */
-void CheckpointPlugin::Render() {
-    if (which_is_being_bound != KEYBIND_ASSIGNWHICH::NONE) {
-        // show the key binding window
-        // creating a translucent background
-        ImGui::SetNextWindowPos(ImVec2(0, 0), ImGuiCond_Always);
-        ImGui::SetNextWindowSize(ImVec2(ImGui::GetIO().DisplaySize.x, ImGui::GetIO().DisplaySize.y));
-        ImGui::PushStyleColor(ImGuiCol_WindowBg, ImVec4(30.f, 30.0f, 30.0f, 0.2f));
-        ImGui::Begin(
-            "translucent_background",
-            NULL,
-            ImGuiWindowFlags_NoDecoration | ImGuiWindowFlags_NoMove | ImGuiWindowFlags_NoSavedSettings
-            | ImGuiWindowFlags_NoBringToFrontOnFocus | ImGuiWindowFlags_NoNavFocus
-            | ImGuiWindowFlags_NoFocusOnAppearing | ImGuiWindowFlags_NoInputs);
-        ImGui::PopStyleColor();
-
-        const static float  scale = 1.4f;
-        static const char* line1 = "Press any key";
-        static const char* line2 = "-- or --";
-        static const char* line3 = "Move an Input Axis";
-        static const ImVec2 linesz1 = ImGui::CalcTextSize(line1) * scale;
-        static const ImVec2 linesz2 = ImGui::CalcTextSize(line2) * scale;
-        static const ImVec2 linesz3 = ImGui::CalcTextSize(line3) * scale;
-
-        ImVec2 totalsz;
-        switch (which_is_being_bound) {
-        case KEYBIND_ASSIGNWHICH::CPT_REWIND_KEY:
-        case KEYBIND_ASSIGNWHICH::CPT_FASTFORWARD_KEY:
-            totalsz += linesz2 + linesz3;
-            [[fallthrough]];
-        default:
-            totalsz += linesz1;
-        }
-        ImVec2 PopupWindowSize = totalsz
-            + ImGui::GetStyle().WindowPadding * 2
-            + ImGui::GetStyle().FramePadding * 2;
-        ImGui::SetNextWindowSize(PopupWindowSize);
-        ImGui::SetNextWindowPos(ImGui::GetIO().DisplaySize * 0.5f, ImGuiCond_Always, ImVec2(0.5f, 0.5f));
-
-        static bool is_open = false;
-        if (!is_open) {
-            ImGui::OpenPopup("Set Keybind##popup");
-        }
-        if (is_open = ImGui::BeginPopupModal(
-            "Set Keybind##popup",
-            NULL,
-            ImGuiWindowFlags_NoDecoration | ImGuiWindowFlags_NoMove | ImGuiWindowFlags_NoSavedSettings)) {
-            ImGui::SetWindowFontScale(scale);
-
-            AlignForWidth(linesz1.x);
-            ImGui::TextUnformatted("Press Any Key");
-            if (which_is_being_bound == KEYBIND_ASSIGNWHICH::CPT_REWIND_KEY || which_is_being_bound == KEYBIND_ASSIGNWHICH::CPT_FASTFORWARD_KEY) {
-                // only show that moving an axis is available for the only two binds it matters for
-                AlignForWidth(linesz2.x);
-                ImGui::TextUnformatted("-- or --");
-                AlignForWidth(linesz3.x);
-                ImGui::TextUnformatted("Move an Input Axis");
-            }
-            ImGui::EndPopup();
-        }
-        ImGui::End();
-    }
-};
-
-/**
- * @brief Returns the name of the menu to refer to it by
- *
- * @return The name used refered to by togglemenu
- */
-std::string CheckpointPlugin::GetMenuName() {
-    return "checkpointplugin";
-};
-
-/**
- * @brief Returns a std::string to show as the title
- *
- * @return The title of the menu
- */
-std::string CheckpointPlugin::GetMenuTitle() {
-    return "CheckpointPlugin";
-};
-
-/**
- * @brief Is it the active overlay(window)?
- *
- * @return true if overlay which isn't interacted with (pluginwindow.h)
- */
-bool CheckpointPlugin::IsActiveOverlay() {
-    return true;
-};
-
-/**
- * @brief Should this block input from the rest of the program?
- * (aka RocketLeague and BakkesMod windows)
- *
- * @return True/False for if bakkesmod should block input
- */
-bool CheckpointPlugin::ShouldBlockInput() {
-    return false;
-};
