+++ version.h
@@ -3,7 +3,7 @@
 #define VERSION_MAJOR 1
 #define VERSION_MINOR 8
 #define VERSION_PATCH 1
-#define VERSION_BUILD 964
+#define VERSION_BUILD 982
 
 #define stringify(a) stringify_(a)
 #define stringify_(a) #a
